backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "patternfly-fastapi-template-backend"
version = "0.1.0"
description = "FastAPI backend for PatternFly FastAPI Template"
authors = [
    {name = "PatternFly FastAPI Template Team"},
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi==0.115.14",
    "uvicorn[standard]==0.34.3",
    "python-dotenv==1.1.1",
    "pydantic-settings==2.10.1",
    # Database dependencies
    "sqlmodel<1.0.0,>=0.0.21",
    "alembic<2.0.0,>=1.12.1",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "tenacity<9.0.0,>=8.2.3",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    # GraphQL
    "strawberry-graphql[fastapi]>=0.245.0",
    "sqladmin>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
